Added tox and flake8 instead of pep8+pyflakes.
* Removed Python 3.2 support temporarily.
