+++ tox.ini
@@ -1,0 +1,25 @@
+[tox]
+skipsdist = True
+envlist =
+    {py27}-django-{18,19,110}
+    {py33}-django-{18}
+    {py34}-django-{18,19,110}
+    {py35}-django-{19,110}
+    {py27,py35}-flake8
+
+[tox:travis]
+3.5 = py35
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}:{toxinidir}
+deps =
+    coverage>=3,<4
+    py27: mock
+    flake8: flake8>=2.0,<3.0
+    django-18: Django>=1.8,<1.9
+    django-19: Django>=1.9,<1.10
+    django-110: Django>=1.10,<1.11
+commands =
+    flake8: flake8 safedelete --ignore=E501
+    django: coverage run --parallel-mode {toxinidir}/runtests.py
